created HeadApp&MainPage
