+++ src/components/HeadApp.vue
@@ -1,0 +1,20 @@
+<template>
+  <div>
+    <h1>The Daily Bugle</h1>
+    <hr />
+  </div>
+</template>
+
+<script>
+
+</script>
+
+<style scoped>
+  @import url("https://fonts.googleapis.com/css2?family=Dancing+Script:wght@400;700&display=swap");
+
+  h1 {
+    font-family: "Dancing Script", sans-serif;
+    margin: 20px;
+    text-align: center;
+  }
+</style>
